split out entity tests
